finished get all spots

+++ backend/db/migrations/20241031214459-create-review.js
@@ -25,7 +25,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       stars: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(2,1)
       },
       createdAt: {
         allowNull: false,
